+++ src/qhttpserverconnection.cpp
@@ -87,7 +87,7 @@ int
 QHttpConnectionPrivate::url(http_parser*, const char* at, size_t length) {
     Q_ASSERT(ilastRequest);
 
-    itempUrl.append(at, length);
+    itempUrl.append(at, static_cast<int>(length));
     return 0;
 }
 
@@ -111,7 +111,7 @@ QHttpConnectionPrivate::headerField(http_parser*, const char* at, size_t length)
         itempHeaderValue.clear();
     }
 
-    itempHeaderField.append(at, length);
+    itempHeaderField.append(at, static_cast<int>(length));
     return 0;
 }
 
@@ -120,7 +120,7 @@ QHttpConnectionPrivate::headerValue(http_parser*, const char* at, size_t length)
     if ( ilastRequest == nullptr )
         return 0;
 
-    itempHeaderValue.append(at, length);
+    itempHeaderValue.append(at, static_cast<int>(length));
     return 0;
 }
 
@@ -197,7 +197,7 @@ QHttpConnectionPrivate::body(http_parser*, const char* at, size_t length) {
         return 0;
     }
 
-    emit ilastRequest->data(QByteArray(at, length));
+    emit ilastRequest->data(QByteArray(at, static_cast<int>(length)));
     return 0;
 }
